backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.1.3
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
elasticsearch>=8.0.0
psycopg2-binary>=2.9.9
sqlalchemy==2.0.48
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov>=4.1.0